// my-app/src/types/product.ts
export interface Product {
    id: number;
    title: string;
    description: string;
    price: number;
    discountPercentage?: number;
    rating?: number;
    stock?: number;
    brand?: string;
    category: string;
    thumbnail: string;
    images: string[];
    liked?: boolean;
  }
  
  export interface Filters {
    category: string;
    showFavorites: boolean;
    searchQuery: string;
    priceMin: string;
    priceMax: string;
  }